Fall through when NewType has the wrong argument number
Summary:
Based on the discussion in the previous diff, I'm relying on the existing logic to create bindings for the LHS for the new type and removing my own check for the number of arguments.

The error messages from before look nice, but this way, we also get the bindings automatically.

Might we worth it to revisit the error messages and see if we can improve them.

I tried avoiding the "unreachable" clause, but `as_slice` is unstable, so I wasn't able to use it to match on the type beforehand. I also considered removing the match statement into the main function, but it seemed that this will cause me to do 1 clone.

Reviewed By: rchen152

Differential Revision: D70263789

fbshipit-source-id: d4b721c244420f8a34214c35701914871642dc87