Make check command take an additional config finder parameter
Summary:
This diff attempts to make `commands::check::Args::run()` take yet another parameter `config_finder`, which is a function that finds the right config file given the path of a checked source file. This is needed to implement the "per-file config file searching" algorithm mentioned in the configuration design doc.

No actual logic is implemented. This diff only sketches the interface.

NOTE: This API currently only works for Meta-internal entry point and the OSS entry point in single-file-check mode. Support for project-check mode for the OSS entry point needs to be added separately.

Reviewed By: stroxler

Differential Revision: D70212965

fbshipit-source-id: 51e737e1084750a71fbfa763c991c5b23ca585f0