Swap the order of override and literal promotion handling
Summary:
The `override` extraction is simpler and has fewer downstream consequences, so it's a little
easier to reason about the code if we do this first.

(Context is that I was considering trying to make `enum` value inferred as a `Literal` but that
led to bugs, and making this change helped me narrow the potential root causes. I gave up on
that behavior change, but I still think this ordering is slightly better).

Reviewed By: migeed-z

Differential Revision: D70266526

fbshipit-source-id: 60d749cf6e8ff81e962ebcd74d2006a0fb2f091d