Expand the comment on enum handling
Summary:
I realized that currently enum value extraction does not use `Literal` types. This is not
obvious from a skim of the code so I thought it was worth a comment. While I'm at it I
added a TODO so we remember that the ideal behavior here is not clear, and I also
summarized what is actually happening in this block since it's nontrivial to skim.

Reviewed By: yangdanny97

Differential Revision: D70266527

fbshipit-source-id: 64233361701b47d755c1bbf88d41869356a6802f