Synthesize an __init__ method for NewType
Summary:
Following up on synthesizing an `__init__` method, we could either create a fake expr or synthesize an `__init__` method in the solving stage.

I looked into both but in the end, I found this solution to be easier since we just need to synthesize a callable type.

There needed to be some extra plubming to indicate that this class was synthesized from a NewType.

In the solving stage, we check the metadata and synthesize the `__init__` method based on the base class, which we first convert to a type.

TODO:
- As seen from the testcase, I think we need to promote literals. I need to think about where exactly this needs to happen though.
- The base class must not be generic. Right now it actually works, but I wonder if it's an accident, so I need to look into it as well.

Reviewed By: rchen152

Differential Revision: D70283719

fbshipit-source-id: b2ebdc3aa0837dd2b74f8c5a07f4f65d534f1856