Add config file flag
Summary:
Adds a `--config-file` flag to Pyre2, which will error if used in check-files mode instead of check-project mode. This path is not yet used, but will eventually enable overriding the default configuration finding behavior.

NOTE: This flag is not yet hooked up to anything. The current plan is to only enable this flag for the `pyre2 check` subcommand, but that is subject to change.

Reviewed By: grievejia

Differential Revision: D70280195

fbshipit-source-id: 30f97094220ea1f0edb63d6189d9ce97bc584c69