Embed Pyre2 executable into conformance test script
Summary: I noticed that the conformance script is invoking Buck under the hood to run the Pyre2 binary. But note that the script is already invoked by Buck from a custom test target. We could easily eliminate the recursive Buck invocation by just having the script itself depend on Pyre2 binary target, similar to how the continuous job is set up. This would in turn eliminate the need of resorting to a different isolation directory (which can be slow to build).

Reviewed By: stroxler

Differential Revision: D70423753

fbshipit-source-id: cc797f81aa6ac67656f71094b55a4fea0fc5d6f0