Change the current test only reduced stdlib to be available for wasm build as well
Summary:
For the demo playground, it's probably a bad idea to include all of the typeshed. It will be too big, and more importantly we currently have trouble compiling the zstd code that processes it into wasm. Therefore, I propose that we should use the reduced set of typeshed that we currently use for tests for now.

This diff factors it out of the module, and make it available for either tests for wasm build.

Reviewed By: dluo

Differential Revision: D70419405

fbshipit-source-id: e7c9ac2f5122ba23ce30401eca32a0acb29ce6e1