Add plumbing to most of `class_field.rs` for simple getter descriptors
Summary:
Continue working our way out in the implementation of descriptors with getters; this diff
adds most of the plumbing for them to `class_field`, so that we call into the `attr.rs` code
in `as_instance_attribute` and `as_class_attribute`.

We don't yet identify descriptors in the calculation to create `ClassField`, so the code is
still not yet live.

Reviewed By: ndmitchell

Differential Revision: D70406724

fbshipit-source-id: a16fb1f4bde0db771d1b0dfc846d926cefdfed49