Make placeholder doc pages public (#134)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/134

We have a lot of issues with broken links; it's tricky to handle moving pages,
especially because if we try to hide placeholder pages, then the internal and
external builds are different and some of the broken link warnings will only
show up on github.

Let's just make all pages public, even if they are still placeholders (the
only remaining placeholders are error-suppressions and getting-started, we
could probably prioritize these soon).

Reviewed By: dluo

Differential Revision: D73594605